test(ui): split overlay renderer fixtures
